Display img and list
An image of the finished dish
Under the image, it should have an appropriately sized “Description”
Under the description, add an “Ingredients” heading followed by an
unordered list of the ingredients needed for the recipe.
Finally, under the ingredients list, add a “Steps” heading followed by
an ordered list of the steps needed for making the dish.